refactor: api 호출 로직과 auth 로직 패키지 분리
서버를 호출하는 api 로직과 auth 관련 resource(supabase, oauth)를 호출하는 auth 로직에 대한 패키지를 분리하였습니다.
