+++ fourtoon-cookie/src/pages/IntroPage/IntroPage.tsx
@@ -10,7 +10,7 @@ import { getMember } from "../../apis/member";
 import GlobalErrorInfoStateContext from "../../components/global/GlobalError/GlobalErrorInfoStateContext";
 import { GlobalErrorInfoType } from "../../types/error";
 import AppleSignInAndSignUpButton from "./AppleSignInAndSignUpButton/AppleSignInAndSignUpButton";
-import { jwtManager } from "../../apis/jwt";
+import { jwtManager } from "../../auth/jwt";
 
 const IntroPage = () => {
     const navigation = useNavigation<NavigationProp<RootStackParamList>>();
